(PUP-7503) Fix hostname mismatch error reporting with Ruby 2.4
Ruby 2.4 introduced a change to OpenSSL that adds a step to
certificate preverification that checks that the hostname matches what
is in the certificate. While this is not enabled by default for most use
cases, any use of `set_params` to configure OpenSSL will enable it.
In the past, the hostname was verified in a post-connection verification
step and produced a specific error message. Now it simply emits a generic
SSL error. Since we still want to give a more informative error message,
this commit changes our error handling to explicitly check for a hostname
mismatch when we receive an SSL error and emit our custom descriptive
message. Note that in Ruby 2.4, this duplicates work already done by Ruby,
but it allows us to be consistent with our error reporting.