(PUP-3552) Centralize Accept and Content-Type validation
Previously, IndirectedRoutes raised 406 Not Acceptable if the Accept header was
missing, while Request raised the same error if the Accept header specified a
format that the server doesn't support.

Also previously we only raised a 415 Unsupported Media Type error if we tried to
read the body of a PUT, and the Content-Type was unsupported. But we may add
code paths in the future that attempt to read the request body, so it's better
to do the validation more centrally in the Request#response_formatter_for
method.

Note we still have to check and possibly raise 415 when reading the body into
the model, because not every model supports every format, e.g. cert-related
models don't support json or binary, only text/plain (aka `s`).

This commit also removes the `Request#format` method, as it's not used in puppet
or puppetserver.

This commit also moves the recently added tests for HTTP 406/415 to
request_spec, so they are not specific to PUT.