(maint) Remove unused AST constant in Scope and move a require
The constant AST was defined but never used. It is now removed.
The 'puppet/parser/resource' was required inside of the Scope class
body. It is now moved outside of the class body.

The require is not really needed anymore, but there are hundreds of
tests that rely on it being required. That is subject for a future
cleanup.