(PUP-7514) Make break() work inside a reduce body
Before this, a call to break() in a reduce lambda would cause an error.
It should act as if the end of input was reached. It did not because
there was no handling of the StopIteration exception.

This commit fixes this problem by catching the StopIteration inside of
the yielding lambda in the reduce function. It was not possible to catch
StopIteration outside of the call to (ruby) reduce in the function as
that does not convey a value to be returned.