(PUP-7381) Use PsychSupport consistently and let to_data_hash return Data
This commit ensures that the `Puppet::Util::PsychSupport` module is
included by all classes that participate in serializaion. It also ensures
that all `#to_data_hash` methods return a hash that is an instance of
the Puppet `Data` type and hence, can be serialized to YAML without
tags being added. With this in place, the YAML and JSON serialization
will produce the same data stream (albeit with diffent syntax).