(PUP-7198)(refactor) separate saving from extracting in compiler
Prior to this commit, the `extract_facts_from_request` method had a side effect
of saving the facts from the request, rather than just extracting them as the
name implied. This in turn meant that the return value for this method was
whatever the facts terminus returned, which could be anything. It also meant
that we never actually locally captured the facts from the request for using
later. This commit splits `extract_facts_from_request` into two methods - the
same extract method that actually only extracts the facts, and a new method,
`save_facts_from_request` that saves the facts from the request. This clears up
the logic, removes the side effect, and allows us to use the facts from the
request later in the compilation.

Signed-off-by: Moses Mendoza <moses@puppet.com>