(PUP-7198)(PUP-3438) use request facts for classification
Prior to this commit, the catalog compiler terminus would extract the facts
from a given compilation request, save them, and then not use them when
requesting classification for a node. Instead the node terminus would look up
the facts for any given request. However, because fact submission can be
asynchronous, it was possible that the facts saved by the compiler terminus
would not ultimately be the ones used by the node terminus when retrieving
classification - possibly instead retrieving stale/outdated facts for a node
despite the correct facts having been submitted with the original compilation
request. This race condition is most evident in environments using exported
resurces with puppetdb that have not deployed (to every compile master) a
routes.yaml which sets the facts cache specifically to yaml instead of
puppetdb. The puppetdb module deploys a routes.yaml file by default, but in a
multi-master environment, the routes.yaml file may only have been deployed to
one of the masters.

This commit updates the catalog compiler terminus to instead thread the facts
object that it extracts from the request through its logic in finding a node,
so that we can ultimately pass this facts object on to the node request. We
extract the facts object first instead of just passing the original request on
because the compiler terminus performs deserialization on the received facts
from the original request and it would be nice to not duplicate that elsewhere.

Signed-off-by: Moses Mendoza <moses@puppet.com>