(PUP-5912) Enable allocate, then initialize pattern on Object type.
In order to create a self referencing type from a hash, it must be
possible to first allocate the instance and then make that allocated
instance available when creating the hash used for initializing the
instance.

This commit adds the needed `#allocate` method and moves initialization
of attributes from the `#initialize` to `#_pcore_init_from_hash`.