(PUP-5912) Add converters to convert value -> data vice versa
This commit adds two converters. One that converts an object into `Data`
value that can be represented using JSON or YAML without loss of type
information and another that can reverse that conversion.

Objects that are not `Data` will be represented as a `Hash` containing
a `__pcore_type__` key. The value of the key is typically a string that
names the type. The remainder of the hash will be either one value
under the key `__pcore_value__` or many values that represent properties
of a complex object.

The `ToDataConverter` is configurable using options. They control what
happens when a `Symbol` is encountered (represent as string with type
loss or hash that describes the symbol), when the same complex object is
encountered more than once (duplicated or referenced), and when custom
types are found (pass full type definition or just the name).