(PUP-7554) Don't error on v3 hiera.yaml in module
A v3 hiera.yaml in the root of a module directory should be a warning,
not an error. It should be roughly equivalent to not having a hiera.yaml
file there at all. This is to allow customers who currently have
hiera.yaml files in modules to install new versions of Puppet without
having to first refactor all their modules and CI systems.