(PUP-7155) Make '%p' string format for a Runtime print quoted #to_s.
There is no way to represent a `Runtime` object as something that Puppet
will parse and recreate an instance from. There are times though, when
it is desirable to output something anyway that does not break the
parser if parsed. This commit therefore defines the programmatic '%p'
representation of a `Runtime` object to be the result of quoting whatever
its #to_s method returns.