(PUP-7155) Integrate PuppetObject #to_s with StringConverter
Before this commit, the only way to output a string formatted `Object`
instance was to directly use its `#to_s` method. That method in turn
would output a nicely formatted string. This commit integrates the
formatting of a `PuppetObject` (done by the `TypeFormatter`) with the
`StringConverter` so that the formatting can be controlled and also
embedded in formatting of hashes and arrays.