(PUP-7155) Make explicit format types override more specialized defaults
Before this commit, an explicitly given format specification hash keyed
with the `Collection` type would be overridden by the default for
more specialized types (i.e. `Array`and `Hash`) and hence, render the
specification useless.

This commit ensures that an explicitly given type overrides a default,
even if the type differ from the default. The rule is instead that
the when default is assignable to the given type, and not equal to the
given type (since that causes a merge of container formats), then it
is overriden.