(PUP-6060) remove superfluous node_cache_terminus setting in lookup
Prior to this commit we would set the :node_cache_terminus setting value to
:write_only_yaml in the lookup application. However, we never set the
indirection cache_class to the value of the setting - which meant regardless of
this setting our cache_class was nil and thus Puppet::Node.indirection would
never find/retrieve from cache. This is fine though. Since we're doing a CLI
based lookup, there isn't a performance-based need to retrieve from cache, and
since lookup is inherently a read-only operation it's actually better not to
have a side-effect of writing node data anyway. It's possible the original code
was just copied from application/master.rb which would explain it being here.

Signed-off-by: Moses Mendoza <moses@puppet.com>