(PUP-7575) Add --job-id flag to `puppet agent`
Commit 4e05e16 added internal handling of job id for agent catalog
requests and report submission, but had no facility to actually provide
the job id to the agent. This commit now adds a --job-id flag which will
simply thread the provided value (which is an arbitrary string) through
to the catalog request and the report. This will allow correlating node
run reports with orchestration jobs in PuppetDB.