(PUP-6517) Raise output and backtrace of rescued exception
Previously if `execfail` method failed to execute the command, it would
try to access the `backtrace` method on the exception class passed into
the method instead of the backtrace of the rescued exception.

This commit modifies the `execfail` to rescue the exception and re-raise its
message and backtrace. It adds missing tests for the execfail method and the
rpm `latest` method.

It also marks the `execfail` and related provider methods as deprecated, since
they are confusing more than anything.