(PUP-6698) Make create_resurces honor @@ to mean exported
Before this, the use of @@ in `create_resources` created resources that
were exported, but not also marked as virtual. The virtual bit is
important since it controls the inclusion in the catalog.

This commit makes the virtual flag be set when exported is set by
`create_resources`.

The existing test did not detect this, since it used a call to `realize`
since it could not otherwise find the resources due to filtering of all
virtual resources. This commit therefore adds a method to compile
without filtering to give tests a chance to check the catalog state
before filtering.