(PUP-1606) Add validation under --strict for virtual/exported class
This adds (static) validation for the constructs:
`@class { ... }` and
`@@class { ... }`
under the control of the --strict flag such that `off` and `warning`
produces a `warning` and `error` flags the expression as an error.

This is put under the `strict` flag as the change woud otherwise be a
breaking change (currently there is no error and the virtual/exported
is simply ignored).