(maint) Make it more convenient to raise an optional runtime issue
Before this it was quite complicated to use an Issue to result in either
an error or a warning from a context without access to the AST node that
caused the problem. While it was possible to raise a SemanticError, that
did not work when the issue potentially was configured to only issue a
warning.

This adds a `runtime_issue` to the Runtime3Support class that takes an
Issue and options and constructs arguments to the rest of the
error/issue reporting system.

To make this work, an instance of SemanticError is used as SourcePos as
the validation diagnostic producer accepts a StandardError that has
file, line, pos as methods as an instance of "semantic". The diagnostic
producer could do with an overhaul as the use of a SemanticError is not
ideal. This implementation is made so to not disrupt the rest of the
issue framework.