(PUP-1795) Place string to numeric coercion under --strict control
Before this strings were automatically and silently coerced to Numeric
when values where required to be Numeric (in arithmetic operations, and
for integer indexes). This was bad because it masks possible data type
and conversion errors.

Now, it is possible to get a warning or error (or ignored) as the
coercion is not under the control of the `--strict` setting.