(maint) Improve hiera.yaml v3 warnings
When a hiera.yaml file version 3 is found at the environment root or
module root Puppet will print a warning, indicating the file will not be
used. Prior to this commit it was not clear exactly which file the
warning pertained to.

This commit adds the specific file to the warning, making it clear what
file is causing the warning to be raised.