(PUP-2280) Mark runs as failed when refresh events fail
Previously, if a 'refreshonly' exec failed, this failure was not
propagated into the report because refresh failures are tracked via a
different resource status, `failed_to_restart` as opposed to just
`failed`.

This commit updates the failure check logic in the transaction to
include checks of `failed_to_restart`, as well as a
similar update to the way the report status is calculated. A run is now
marked as failed if any of its resources are marked `failed_to_restart`.