(docs) correct code comment for Indirection#prepare
The comment said this method prepares a request, which is not true. In fact,
this method never prepared a request - it was added in 2008 in f9881edd, but
even then it only selected the terminus and checked authorization. This commit
corrects the message and adds param/return doc strings so at least followers can
tell at a glance what's happening here.

Signed-off-by: Moses Mendoza <moses@puppet.com>