(PUP-7413) remove check for `:select_terminus` in indirection
The last class in Puppet that defined `:select_terminus` was removed in 40ee670
in Puppet 2.7.18, but this check has remained in Indirection.rb ever since. The
method it is in is defined as api private, and it's now just cruft. Remove it
because its superfluous. A github search didn't show code that actually defined
this method, except for repos that vendored an old version of puppet.

Signed-off-by: Moses Mendoza <moses@puppet.com>