(PUP-7508) Add explain output or mapped_paths undef/empty mapped var
This adds explain output for the two cases when mapped_paths is used and
the mapped variable is either `undef` or empty. This is helps a user
identify if paths are not used because of missing fact/variable, empty
fact input or because of mapped result does not reference existing
files.

The test is currently failing.