(PUP-3042) Make logged errors in report set status == failed
Before this, a circular dependency would not alter the state of a report
for an agent run. This is bad because users are then not alerted to the
problem and believe that the catalog application was correct.

Now the report's status is set to 'failed' if there is any error
included in the logs of the report.

Logs are given to the report both before it is finalized and after. This
requires a check when computing the status at the time the report is
finalized, and when any additional runtime log entries are added to the
report.

A test is added for the specific error of circular dependencies. The
status will however be set to 'failed' for any logged error.

This alters two tests that assumed that it was ok to call << with a
string where it should have been an instance of Puppet::Util::Log. (The
`<<` method is api private.)