(PUP-7198)(PUP-3438) accept facts in Puppet::Node#fact_merge
Prior to this commit, calling #fact_merge on a Puppet::Node would always result
in a Puppet::Node::Facts.indirection#find call - ie the node's facts would be
queried via the indirector and then merged with the node's parameters. This
commit updates #fact_merge to accept an optional facts parameter representing
the facts to merge. This is part of a larger effort to use the same facts
supplied by a node all the way through a node find operation and minimize
needless calls to the indirector (which often result in calls over the wire).

Signed-off-by: Moses Mendoza <moses@puppet.com>