(PUP-7260) Disable the audit metaparameter
The audit metaparameter was deprecated in Puppet 4.10. This commit
disables it, so that while it is still allowed to be present in
manifests, it will do nothing.

This commit removes the mechanics of auditing, as well as references
throughout the codebase and in the tests. It also updates some details
of the report schema to stop referencing audit.