(maint) Optimize #instance? method on PObjectType
This commit optimizes how a PObjectType can determine if a value is an
instance of the given type. Instead of performing a, potentially very
expensive, `TypeCalculator#infer_set` on the type, a new streamlined
method is added that checks if the value is a `PuppetObject` or a class
that is registered with the implementation registry.

The real benefit kicks in when the given value is a complex and large
object that is not an instance of the type.