(PUP-5912) Make Catalog use ToDataConverter/FromDataConverter
This commit changes the way a Resource containing parameters with rich
data is serialized so that it takes full advantage of the serialization
semantics implemented in the `ToDataConverter` and `FromDataConverter`
classes.

The conversion controlled by the setting of the `rich_data` flag. If
not set, then rich data values will be serialized as strings and warnings
will be printed. This behavior is unchanged but the actual management
of such values has been moved into the `ToDataConverter`.