(PUP-7603) Make StringFormatter use single quotes more diligently
Prior to this commit, the StringFormatter would use double quotes for
all strings that contained control-, non-ascii-, dollar, or backslash
characters. Double quoting is however only needed for strings that
contain control- or non-ascii characters.

This commit changes how the detection is made so that single quotes are
used whenever possible.