(PUP-7382) Ensure that Transaction::Event is serializable
The `Transaction::Event` class contains _previous_, _desired_, and
_historical_ values that stems from the _current_ value of a provider.
These values in turn, may not be serializable using JSON or YAML since
they might not conform to the `Data` type.

This commit ensures that the hash returned by the `Event#to_data_hash`
method really is `Data` compliant by passing it through the
`ToDataConverter#convert` method. This means that the range of JSON/YAML
serializable values is extended from just being `Data` to instead be
`RichData`. It also means that instances of classes, unknown to the
puppet type system, will be serialized as strings, formatted using the
`StringConverter` which in turn will emit the same string as will be
produced by the `Parameter#format_value_for_display` method.