(PUP-7381) Remove test asserting that a vertex can reference its graph
This commit removes a test that asserted that an arbitrary object could
be used as a vertex, and that this vertex then could reference the graph
that contained it.

The reason for removal was that the only way such a graph can do a
roundtrip through serialization, is by using YAML and YAML specific tags.

When dropping the old `#to_yaml_properties` approach in favor of using
`#encode_with`, the YAML specific tags for backreferences does not work
anymore for Ruby versions < 2.3.

Since there is no known use-case where such a roundtrip is needed, and
since we want to ensure that a graph can be represented as `Data`,
there's little reason to spend time and resources making the test work.