(PUP-7608) Ensure that a single array is "merged" during lookup
This commit changes the merge logic so that even a single entry is
subjected to a merge strategy. The change is needed to ensure that an
array containing duplicates is made unique when the unique merge
strategy is used and that this happens regardless of if it is merged
with other arrays or not.