(PUP-7543) Ensure that the eyaml_lookup_key gets configured on each call
Prior to this commit, the configuration of the eyaml backend was only
performed once per entry. That didn't work when several different
configurations were used, since the config is stored statically in an
Options class. This commit ensures that the configuration is instead
performed each time an encrypted entry is decrypted.