(PUP-6359) Fail fileserver auth for mounts with non-global allow
The use of security directives in fileserver.conf mount points has been
deprecated for multiple releases in favor of the use of the auth.conf file.
This commit is a step toward complete removal of support for these
directives.

With this commit, requests will be denied access to a mount point at the
fileserver indirector layer, with a corresponding error message logged, if any
non-global allow (allow *) directives are present in the mount point definition.

Previously, requests made to mount points for which no security directives were
defined would fail authentication at the fileserver indirector layer.  With
this commit, these requests would be granted authentication at the fileserver
indirector layer.