(MAINT) Explicitly configure autosign to false for dup-csr test
The ticket_3360_allow_duplicate_csr_with_option_set.rb test contains a
couple of steps which sign certs via the puppet cert cli.  If the test
were run with the `autosign` setting in the puppet.conf set to `true`,
the sign steps could "fail" because they end up being no-ops since CSRs
are signed immediately as they are submitted.  This commit sets
`autosign` to `false` while the test is running in order to ensure that
the CSRs have not already been signed before the cli sign steps in the
test are run.