(PUP-7617) Only use Windows locale path on Windows
Because the config file for locale tooling is a relative path, in some
situations the Windows path is a valid path on Unix systems, even though
it does not point to the right location. This causes Puppet to look for
the locales config file in the wrong place and fail. This commit adds a
check, so we only try to use the Windows path on Windows, even if it is
valid elsewhere. It also adds the opposite check for the posix file
system path.