(PUP-7605) Treat 'main' as a reserved class name
This commit adds logic to prevent that `Class[Main]` is ever created.
There is one exception, and that is when a class is created with an
empty string. This only happens when the Compiler creates its main
resource.

All attempts to create class resource using the symbol :main or the
string 'main' will henceforth yield a "'main' is a reserved class name"
error.

Not being able to create a Class[main] resulted in an error during the
call to `Resource#to_ral`. The type created a component which in turn
recreated the resource that the type originated from. This problem was
fixed by letting the component reference the original resource instead
of recreating it with all bells and whistles.