(PUP-7255) Error gracefully if we fail to send report
If the server version is less than 5, we fail to submit our report, and
we're not already using PSON, then generate a meaningful error message.

    Error: Could not send report: Server version 4.10.1 does not accept reports in 'json', use `preferred_serialization_format=pson`

This case should only ever happen if the agent is running off of a cached
catalog, and it never retrieves file_metadata before trying to send the
report.

Previously, we received an HTTP 500 error indicating the mime type
wasn't supported, because 4.x servers don't correctly return HTTP 415.