(PUP-7620) Add the `unique` function to puppet
Before this, the `unique` function required `stdlib` to be present and
the version in `stdlib` only handles `String` and `Array` values and
does not provide any means to affect how the uniqueness is computed.
(Most notably not being able to do case independent compares for
strings).

The new version handles all `Iterable` values and supports the use of a
lambda to compute the value used for uniqueness computation.

For full details of the function see the yard documentation in the
function itself.