(PUP-3940) Exclude dot from the agent's Accept header
Previously, the agent included `text/dot` in its list of accepted content types.
However, if the server ever returned a catalog in that format, the agent would
error with:

    Puppet::Resource::Catalog does not respond to from_dot; can not intern instances from text/dot

The issue is that `dot` is not a network serialization format. It's a one-way
transformation from the catalog into a graph representation.

This commit ensures `dot` is excluded from the agent's Accept header for all
model types, not just catalogs.

Note this commit doesn't prevent the server from returning a catalog in dot, as
you might want to curl the server to export the catalog:

    curl -H 'Accept: text/dot' https://.../puppet/v3/catalog/<name>?environment=<env>