(PUP-7555) Doc removal of auth support in default fileserver.conf file
This commit adds some documentation to the default fileserver.conf file
which indicates that, per PUP-6359, the use of auth directives is no
longer supported in the file.  The commit expands the example showing
how to implement auth directives in the auth.conf file instead of
fileserver.conf to include info about the format that Puppet Server's
HOCON-based auth.conf file uses.