(PUP-7623) Ignore override with default value
Before this, if an override was performed and the resource had a value,
an attempt would be made to override it with a default value. That
resulted in an error.

This behavior changed when the evaluation and setting of default values
was made earlier in the process to allow collections to be based on
default values being set. As a consequence we ended up with the new
behavior where an attempt is made to override a set value with a default
one.

This commit fixes the problem by simply checking if an override is made
with a default value, and if so simply ignore this.