(PUP-7382) Fix typo in ToDataConverter and add more tests
This commit fixes a typo (is_a instead of is_a?) in the `ToDataConverter`
class and adds test to cover some areas of the `ToDataConverter` /
`FromDataConverter` code that was not tested.

A new `Issue` is also added to improve the message for the case when a hash
key must be converted to a `String`.