(PUP-7642) Allow rich-data of unknown type in reports
Before this commit, an attempt to send a report containing custom rich
data types (Objects) to a server would result in an "No implementation
mapping found for Puppet Type xxx" error. Since there's really no reason
why such types should be internalized in a report, this is now changed
so that it instead keeps the original rich-data hash.