(PUP-7255) Downgrade to pson for older servers
If we make a REST request, and the server's major version, as indicated
by the X-PUPPET-VERSION header, is less than 5, then fallback to pson
for future requests, and emit a warning:

    Warning: Downgrading to PSON for future requests

Because we downgrade the preferred serialization format, we can
successfully send the report later.

Do nothing if we're already configured to use pson or if the header is
missing.