(PUP-7650) Ensure that Loader::TypedName always uses lower-case
Before this commit, if an instance of `Puppet::Pops::Loader::TypedName`
was created from a name that contained upper case letters, those letters
would be preserved. This meant that the responsibility for downcasing
the name was on the caller.

This commit removes the `downcase` from all calls to the `TypedName`
constructor and instead adds a `downcase` inside the `#initialize`
method.