(PUP-7652) Allow quotes in report yaml uuid value for catalog_uuid test
Previously, the catalog_uuid_correlates_catalogs_with_reports.rb test
expected the value that it finds for the catalog_uuid in a report yaml
file for an agent run would not be surrounded with single quotes.
Starting with Ruby 2.3.2, however, some uuid values may be surrounded
with single quotes.  Quoting around the value is legal for the yaml
generator to do in either case.  This commit changes the test to
optionally allow for quotes around the value so that it will pass
reliably across Ruby versions.