(maint) correct comment in CharacterEncoding::scrub
An early iteration of the scrub method always replaced invalid sequences with
'?'. This was amended to use the unicode escape character when the string is
UTF-8 or UTF-16LE, and only revert to '?' if the string is not in one of those
encodings. However, the doc string for the method was not updated. This commit
corrects it.

Signed-off-by: Moses Mendoza <moses@puppet.com>