(PUP-7602) Only support binary data in JSON for catalogs
Prior to commit a7b14ba9, the base JSON terminus supported binary data
because it used PSON for serialization. Commit a7b14ba9, switched it to
using JSON instead, which had the side-effect of no longer supporting
binary data in JSON documents.

The one place where we want to support binary data is in the catalog,
such as a file resource with a `content` parameter, see PUP-5584.

This commit modifies the base JSON terminus so that it performs file I/O
in binary mode, so that ruby does not transcode the contents to/from the
default external encoding. But when converting between the textual
string and JSON, it will continue to require valid UTF-8.

To support the binary catalog use-case, this commit modifies the
catalog-specific JSON terminus so that it can fallback to PSON if JSON
conversion fails. The fallback case will occur if a catalog contains
binary content and:

 * The server attempts to serialize to JSON in the HTTP response
 * The agent attempts to serialize its cached catalog
 * The agent attempts to deserialize its cached catalog

The fallback case won't happen when an agent deserializes the HTTP
response from the server, because the server will send `Content-Type:
text/pson`, so the agent will know to expect PSON.

We also use Encoding::BINARY instead of ASCII_8BIT as the former is
preferred in puppet's codebase.