(PUP-7602) Try each supported format during find and search
Previously, find and search REST APIs would always attempt to render the
object(s) using the first format that the client accepts and server supports.

The delete and save methods are unchanged, because they do not render
meaningful data in the HTTP response.

If a catalog contains binary content, then attempting to serialize to JSON will
fail with a FormatError. If that happens retry using msgpack (if
enabled), or PSON.