(PUP-7660) Skip relationship validation of virtual resources
Before this virtual resources were included in the validation of
relationships. This could cause problems since the reason a resource is
still virtual by the time the catalog is validated means that it simply
should not be used, and it is therefore likely that it has references to
things that are not present.