(PUP-7671) Ensure that TypeCollection uses downcased name for load
Before this commit, if the `TypeCollection#find_or_load` was called with
uppercase characters in the name, it would make an attempt to first find
a resource using a munged (downcased) name, and then, if the resource
was not found, load the resource using the name without downcasing it.

This commit ensures that both find and load uses a downcased name.