(PUP-5893) Make it possible to create a Hash tree
This adds the ability to create a Hash-tree from an array created by a
tree_each iteration. The added feature is to accept an optional Boolean
flag to Hash.new that when set to true means that a hash-tree should be
built. In this mode, tuples consisting of [path, value] entries are used
where path is an array with 0 or more keys. The tree is populated with
hashes unless container nodes were included.