(PUP-7277) Remove type_parser argument to PAnyType#resolve method
When the TypeParser was first designed, it contained the scope necessary
to load types during parsing. The parser was therefor a necessary
parameter to the `PTypeAny#resovle` method with potentially would need
to call the same parser recursively. This is no longer the case. Not
only was the scope replaced by a loader, it was also broken out as a
separate argument so that the `TypeParser` actually became a fully
singleton class.

This commit therefore removes the now redundant `type_parser` argument
in all the `#resolve` method and its support methods, and also to
the `DispatchBuilder` used when building a `Function`.