(PUP-5893) Add option to Hash new to create Hash from flat tree
This adds the ability to create a Hash from a "flat tree", thereby again
turning into a tree. By giving a second argument to Hash.new of `tree`
or `hash_tree` an `Array` is interpreted as having tuple entries on the
form [[path, val], [path, val]], where path is an array of indexes from
the root to the value.

The option `hash_tree` produces hashes for all array values. The option
`tree` transforms a top level array to Hash while preserving array/hash
content in general.