(PUP-7277) Add new Init[T] type to the Puppet Type System
This commit adds the type `Init[T]`. It represents all valid initializers
that can be used when creating instances of the type `T`.

Without arguments, the type represents all instances of `RichData`.

The first argument, when given, must be a type. Subsequent arguments
can be used to further qualify the created type. When such arguments are
used, the complete argument tuple used when creating a type will consist
of a given value, followed by those arguments. E.g.

$x = Init[Integer,16].new('80')

will essentially be the same as doing:

$x = Integer.new('80', 16)

(both creates the Integer 128).

The Init type will initialize itself using the disptachers found in the
Function class returned from the #new_function of the given type which
means that using `Init[T]` with a `T` that cannot be instantiated will
raise an exception.