(PUP-6256) Ensure that agent run returns exit code instead of log entry
Before this commit, the `Agent#run` method would pass a block to the
`Agent#run_in_fork` that, in case a `StandardError` was raised, would
log the exception and then return the created `Log` instance. Since the
caller of the block expects an exit code, which it passes on in a call
to exit, it would fail with another exception when it instead passed
the returned `Log` instance.

This commit ensures that the block instead returns the integer 1 in case
a `StandardError` is raised.