(maint) Update comments in commit Rake task
Without this patch applied, build failures caused by commit summaries
that fail to match patterns as specified in `Rakefile` are providing
instructions not in line with actual practice in this project.

The Jira ticket is usually given as PUP-<digits>; doc updates are given
usually as (docs) and occasionally as (docs)(DOC-<digits>).

This patch updates the hint to give instructions in line with the actual
practice in the project.