(PUP-7691) Ensure that the 4x loader is present when epp face executes
Before this commit, the command 'puppet epp' would fail when the epp-
template contained code that required the 4x loader. The reason was that
the compiler was not properly initialized with its overrides. This
commit ensures that the compiler adds the overrides just as it would
when doing a proper compile of a catalog.