(PUP-6290) Ignore, warn, or fail when evaluated resources are overridden
This commit changes the `Puppet::Parser::Resource#set_parameter` so that
it checks if the resource has already been evaluated. If that's the case,
then the behavior is controlled by the setting of `Puppet[:strict]`:

strict=off, the attempt silently sets the parameter although no
re-evaluation of the resource takes place. In essence, the new value is
ignored (same as before this commit).

strict=warning, same as strict=off, but the attempt is also logged as a
warning.

strict=error, compilation fails.